separate utils into individual imports; org utils directory
