+++ src/utils/arrays/index.js
@@ -1,0 +1,4 @@
+import { each } from './each';
+import { indexOf } from './indexOf';
+
+export { each, indexOf };
