+++ src/utils/dom/index.js
@@ -1,0 +1,5 @@
+import { removeNode } from './removeNode';
+import { replaceNode } from './replaceNode';
+import { documentFragments } from './documentFragments';
+
+export { documentFragments, removeNode, replaceNode };
