+++ src/utils/dom/removeNode.js
@@ -1,0 +1,7 @@
+/**
+ * removes a given element in dom
+ * @param {HTMLElement}
+ */
+export const removeNode = (node) => {
+  node.parentElement.removeChild(node);
+};
