+++ src/utils/is/index.js
@@ -1,0 +1,29 @@
+import { isElementAttribute } from './isElementAttribute';
+import { isEqual } from './isEqual';
+import { isEvent } from './isEvent';
+import { isHtml } from './isHtml';
+import { isNumeric } from './isNumeric';
+import { isVNode } from './isVnode';
+import { isWindow } from './isWindow';
+
+import { isArray, isBoolean, isError, isFunction, isNull, isNumber, isObject, isRegExp, isString, isUndefined } from './common';
+
+export {
+  isArray,
+  isBoolean,
+  isError,
+  isEqual,
+  isFunction,
+  isNull,
+  isNumber,
+  isObject,
+  isRegExp,
+  isString,
+  isUndefined,
+  isElementAttribute,
+  isEvent,
+  isHtml,
+  isNumeric,
+  isVNode,
+  isWindow
+};
