+++ src/utils/objects/clone.js
@@ -1,0 +1,7 @@
+/**
+ * shallow copies an object
+ * @private
+ * @param {Object} obj
+ * @return {Object}
+ */
+export const cloneObject = (obj) => Object.assign({}, obj);
