+++ src/utils/objects/index.js
@@ -1,0 +1,5 @@
+import { cloneObject } from './clone';
+import { deepCloneObject } from './deepClone';
+import { iterate } from './iterate';
+
+export { cloneObject, deepCloneObject, iterate };
