+++ src/utils/strings/index.js
@@ -1,0 +1,4 @@
+import { convertStringValue } from './convertStringValue';
+import { startsWith } from './startsWith';
+
+export { convertStringValue, startsWith };
